PR comments * Change log correction * Move test to resolver package
